Refactor BenchmarkModel for argument simplicity
Moving build and run methods inside the base class and making the
derived classes only handling the arguments, defaults etc.

Also making sure all parameters are indeed passed down to the
benchmarks, from compiler and linker options, to benchmark execution.

Finally, moving the size argument is now passed down as is and it's up
to the model to convert that into flags (that can be overwriten by the
user), insdead of intrinsic functionality.

Tested with both benchmarks on x86_64 passing all flags and changing
size and iterations.